Create database fixture to be used in all tests that require a database and implement it in all tests
